#1815 Added toggle to export data. Made new project with GTAS share objects. Refactored code to seperate gtas-commons from gtas-share. Added auditing fields to hit maker.
