+++ gtas-parent/gtas-commons/pom.xml
@@ -28,7 +28,7 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<elasticsearch.version>7.2.0</elasticsearch.version>
 		<hazelcast.version>3.11</hazelcast.version>
-
+		<share.version>0.1-SNAPSHOT</share.version>
 		<hibernate.connection.driver_class>org.mariadb.jdbc.Driver</hibernate.connection.driver_class>
 		<hibernate.dialect>org.hibernate.dialect.MySQL5InnoDBDialect</hibernate.dialect>
 		<hibernate.connection.username>root</hibernate.connection.username>
@@ -36,6 +36,11 @@
 
 	</properties>
 	<dependencies>
+		<dependency>
+			<groupId>gov.gtas</groupId>
+			<artifactId>gtas-share</artifactId>
+			<version>0.1-SNAPSHOT</version>
+		</dependency>
 		<!-- commons -->
 		<dependency>
 			<groupId>org.apache.commons</groupId>
@@ -337,7 +342,7 @@
 			<artifactId>mockito-core</artifactId>
 			<scope>test</scope>
 		</dependency>
-		
+
 		<dependency>
     		<groupId>com.github.dhorions</groupId>
    			 <artifactId>boxable</artifactId>
@@ -349,7 +354,6 @@
     			<artifactId>pdfbox</artifactId>
     			<version>2.0.17</version>
 		</dependency>
-		
 	</dependencies>
 	<build>
 		<plugins>
