+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/model/HitMaker.java
@@ -18,7 +18,7 @@ import java.util.Set;
 @Entity
 @Table(name = "hit_maker")
 @Inheritance(strategy = InheritanceType.JOINED)
-public abstract class HitMaker extends BaseEntity {
+public abstract class HitMaker extends BaseEntityAudit {
 
 	@Column(name = "hm_hit_type", nullable = false)
 	@Enumerated(EnumType.STRING)
