+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/services/jms/AdditionalProcessingMessageSender.java
@@ -1,4 +1,4 @@
-package gov.gtas.summary.jms;
+package gov.gtas.services.jms;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
