+++ gtas-parent/gtas-job-scheduler-war/src/main/java/gov/gtas/job/scheduler/LoaderMessageReceiver.java
@@ -9,7 +9,7 @@ import javax.jms.Session;
 
 import gov.gtas.summary.EventIdentifier;
 import gov.gtas.summary.MessageAction;
-import gov.gtas.summary.jms.AdditionalProcessingMessageSender;
+import gov.gtas.services.jms.AdditionalProcessingMessageSender;
 import gov.gtas.model.MessageStatus;
 import gov.gtas.model.MessageStatusEnum;
 import gov.gtas.model.Pnr;
