+++ gtas-parent/gtas-share/src/main/java/gov/gtas/summary/MessageAction.java
@@ -15,6 +15,7 @@ public enum MessageAction {
     RAW("RAW"),
     PROCESSED_RAW("PROCESSED_RAW"),
     PROCESSED_PNR("PROCESSED_PNR"),
+    PROCESSED_MESSAGE("PROCESSED_MESSAGE"),
     PROCESSED_APIS("PROCESSED_APIS"),
     HIT("HIT"),
     ERROR("ERROR"),
