+++ gtas-parent/gtas-share/src/main/java/gov/gtas/summary/MessageAddress.java
@@ -1,8 +1,5 @@
 package gov.gtas.summary;
 
-import gov.gtas.model.Address;
-import org.springframework.beans.BeanUtils;
-
 public class MessageAddress {
     private String line1;
 
@@ -21,20 +18,6 @@ public class MessageAddress {
     private String flightIdTag;
 
     private String messageHash;
-
-    public static MessageAddress from(String flightIdTag, String messageHash, Address address) {
-        MessageAddress pa = new MessageAddress();
-        pa.setCity(address.getCity());
-        pa.setCountry(address.getCountry());
-        pa.setLine1(address.getLine1());
-        pa.setLine2(address.getLine2());
-        pa.setLine3(address.getLine3());
-        pa.setMessageHash(messageHash);
-        pa.setState(address.getState());
-        pa.setFlightIdTag(flightIdTag);
-        pa.setMessageHash(messageHash);
-        return pa;
-    }
 
     public String getLine1() {
         return line1;
