+++ gtas-parent/gtas-share/src/main/java/gov/gtas/summary/MessagePhone.java
@@ -1,7 +1,5 @@
 package gov.gtas.summary;
 
-import gov.gtas.model.Phone;
-import org.springframework.beans.BeanUtils;
 
 public class MessagePhone {
     private String number;
@@ -9,14 +7,6 @@ public class MessagePhone {
     private String flightIdTag;
 
     private String messageIdTag;
-
-    public static MessagePhone from(String messageIdTag, String flightIdTag, Phone phone){
-        MessagePhone passengerPhone = new MessagePhone();
-        passengerPhone.setNumber(phone.getNumber());
-        passengerPhone.setMessageIdTag(messageIdTag);
-        passengerPhone.setFlightIdTag(flightIdTag);
-        return passengerPhone;
-    }
 
     public String getNumber() {
         return number;
