+++ gtas-parent/gtas-share/src/main/java/gov/gtas/summary/MessageSummaryList.java
@@ -3,7 +3,7 @@ package gov.gtas.summary;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MessageSummaryList {
+public class  MessageSummaryList {
 
     private EventIdentifier eventIdentifier;
     private MessageAction messageAction;
