+++ gtas-parent/gtas-share/src/main/java/gov/gtas/summary/PassengerIds.java
@@ -1,8 +1,6 @@
 package gov.gtas.summary;
 
 
-import gov.gtas.model.PassengerIDTag;
-import org.springframework.beans.BeanUtils;
 
 public class PassengerIds {
 
@@ -11,17 +9,6 @@ public class PassengerIds {
     private String tamrId;
 
     private Long pax_id;
-
-    public static PassengerIds from(PassengerIDTag passengerIDTag) {
-        PassengerIds pids = new PassengerIds();
-        if (passengerIDTag == null) {
-            return pids;
-        }
-        pids.setIdTag(passengerIDTag.getIdTag());
-        pids.setPax_id(passengerIDTag.getPax_id());
-        pids.setTamrId(passengerIDTag.getTamrId());
-        return pids;
-    }
 
     public String getIdTag() {
         return idTag;
