+++ gtas-parent/gtas-share/src/main/java/gov/gtas/summary/PassengerTrip.java
@@ -1,7 +1,5 @@
 package gov.gtas.summary;
 
-import gov.gtas.model.PassengerTripDetails;
-import org.springframework.beans.BeanUtils;
 
 public class PassengerTrip {
 
@@ -25,20 +23,7 @@ public class PassengerTrip {
 
     private Integer hoursBeforeTakeOff;
 
-    public static PassengerTrip from(PassengerTripDetails ptd) {
-        PassengerTrip pt = new PassengerTrip();
-        pt.setCoTravelerCount(ptd.getCoTravelerCount());
-        pt.setDebarkation(ptd.getDebarkation());
-        pt.setDebarkCountry(ptd.getDebarkCountry());
-        pt.setEmbarkation(ptd.getEmbarkation());
-        pt.setEmbarkCountry(ptd.getEmbarkCountry());
-        pt.setHoursBeforeTakeOff(ptd.getHoursBeforeTakeOff());
-        pt.setPassengerId(ptd.getPassengerId());
-        pt.setNumberOfDaysVisaValid(ptd.getNumberOfDaysVisaValid());
-        pt.setPnrReservationReferenceNumber(ptd.getReservationReferenceNumber());
-        pt.setReservationReferenceNumber(ptd.getReservationReferenceNumber());
-        return pt;
-    }
+
 
     public Long getPassengerId() {
         return passengerId;
