+++ gtas-parent/pom.xml
@@ -204,7 +204,8 @@
 				<activeByDefault>true</activeByDefault>
 			</activation>
 			<modules>
-		        <module>gtas-parsers</module> 
+                <module>gtas-share</module>
+                <module>gtas-parsers</module>
 		        <module>gtas-loader</module> 
 		        <module>gtas-commons</module>
 		        <module>gtas-rulesvc</module>
