+++ gtas-parent/scripts/db/1.13.0_to_1.14.0/hitmaker_update.sql
@@ -1,0 +1,17 @@
+alter table hit_maker
+    add
+    created_at datetime null;
+
+alter table hit_maker
+    add
+        created_by varchar(20) null;
+
+alter table hit_maker
+    add
+        updated_at datetime null;
+
+alter table hit_maker
+    add
+        updated_by varchar(20) null;
+
+
